Feature/octree node enum (#28)
* Improve performance of octree traversal

* Inline Vec3 methods

* Change default octree parameters for meshfile components

* Remove unused method

* Inline vec3 methods

* Formatting

* Update version and changelog

* Update octree to use an enum for nodes

* Update octree useage

* Default constructor for Octree

* Format

* Update changelog

* Changelog

+++ src/types/computation/functions/mesh_file.rs
@@ -16,7 +16,7 @@ use crate::{
 };
 
 const MAX_LEAF_TRIANGLE_COUNT: usize = 24;
-const MAX_TREE_DEPTH: u32 = 12;
+const MAX_TREE_DEPTH: usize = 12;
 
 /// Distance function from a mesh loaded from a file.
 ///
